scan

+++ proto/cs/scan/scan.go
@@ -83,7 +83,7 @@ func SendScannerReq(conn net.Conn, table string, slots []byte, store int, fields
 		Store:   int32(store),
 		Fields:  fields,
 		All:     allfields,
-		Timeout: int32(deadline.Sub(time.Now())),
+		Timeout: int64(deadline.Sub(time.Now())),
 	}, deadline)
 }
 
@@ -100,7 +100,7 @@ func SendScanNextResp(conn net.Conn, errCode int, slot int, rows []*flyproto.Row
 }
 
 func SendScanNextReq(conn net.Conn, count int, deadline time.Time) error {
-	return cs.Send(conn, &flyproto.ScanNextReq{Count: int32(count), Timeout: int32(deadline.Sub(time.Now()))}, deadline)
+	return cs.Send(conn, &flyproto.ScanNextReq{Count: int32(count), Timeout: int64(deadline.Sub(time.Now()))}, deadline)
 }
 
 func RecvScannerReq(conn net.Conn, deadline time.Time) (*flyproto.ScannerReq, error) {
